Update game and utility functionalities
This commit contains a series of code changes targeting the game and utility functionalities. The anti-recoil feature was added, some update entity list actions were moved into a condition for the ESP configuration flag, and checks for 'InBuyZone' were added. Also, a number of new read/write functions were added in the memory class and an update was made to the aimbot settings. The CS2Internal offsets were updated to reflect the new memory addresses of various game functions.